+++ src/yrt_input_validator.mjs
